% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatom_df.R
\docType{data}
\name{diatom_df}
\alias{diatom_df}
\title{Diatom df}
\format{
A data frame with 100 rows and 4 columns
\describe{
\item{p_conc}{Phosphate (PO4) concentration (mg/L)}
\item{si_conc}{Silicate (SiO2) concentration (mg/L)}
\item{temp}{Lake water temperature (C)}
\item{predominant_diatom}{Output of the predominant_diatom() function}
}
}
\usage{
diatom_df
}
\description{
Data testing the function predominant_diatom
}
\author{
Mariano Viz
}
\keyword{datasets}
