% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_summary.R
\name{fish_summary}
\alias{fish_summary}
\title{fish_summary}
\usage{
fish_summary(name_vec, graph = FALSE)
}
\arguments{
\item{name_vec}{vector of fish names}

\item{graph}{histogram with count of fish types (default = FALSE)}
}
\value{
Total number of fish = t, Most common fish = c, Rarest fish = l, Histogram with count of fish types = p
}
\description{
The function takes a vector and find the most common fish, the rarest fish, the total number of fish, and has the option to plot a histogram of the number of each fish type.
}
\author{
Claudia Flores & Mariano Viz
}
