% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_annual_discharge.R
\docType{data}
\name{max_annual_discharge}
\alias{max_annual_discharge}
\title{max_annual_discharge}
\format{
A data frame with 10 rows and 4 columns
\describe{
  \item{R}{ranking value}
  \item{max_annual_discharge_m3_s}{maximum annual discharge (m^3/s)}
  \item{year}{year when maximum annual discharge value was recorded}
  \item{Y}{the total number of records for maximum annual discharge in the data frame}
}
}
\usage{
max_annual_discharge
}
\description{
data for recurrence_interval_prob function
}
\author{
Claudia Flores
}
\keyword{datasets}
