% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predominant_diatom.R
\name{predominant_diatom}
\alias{predominant_diatom}
\title{predominant_diatom}
\usage{
predominant_diatom(
  p_contentration,
  si_concentration,
  lake_temp,
  p_min_ast = 0.01,
  p_min_cyc = 0.2,
  si_min_ast = 1.9,
  si_min_cyc = 0.6,
  min_lake_temp = 4,
  max_lake_temp = 28
)
}
\arguments{
\item{si_concentration}{silicate (SiO2) concentration (mg/L)}

\item{lake_temp}{lake water temperature (C)}

\item{p_min_ast}{minimum amount of phosphate needed for Asterionella to have a positive rate of growth (mg/L) (default = 0.01)}

\item{p_min_cyc}{minimum amount of phosphate needed for Cyclotella to have a positive rate of growth (mg/L) (default = 0.2)}

\item{si_min_ast}{minimum amount of silicate needed for Asterionella to have a positive rate of growth (mg/L) (default = 1.9)}

\item{si_min_cyc}{minimum amount of silicate needed for Cyclotella to have a positive rate of growth (mg/L) (default = 0.6)}

\item{min_lake_temp}{minimum lake water temperature for diatoms to survive (C) (default = 4)}

\item{max_lake_temp}{maximum lake water temperature for diatoms to survive (C) (default = 28)}

\item{p_concentration}{phosphate (PO4) concentration (mg/L)}
}
\value{
cyc_predominates, ast_predominates, coexistence, no_survival_conditions
}
\description{
This function determines the predominance between the diatoms Asterionella formosa and Cyclotella meneghiniana in lake waters as a function of the concentration of limiting nutrients phosphate (PO4) and silicate (SiO2).
}
\author{
Mariano Viz
}
