% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_interval_prob.R
\name{recurrence_interval_prob}
\alias{recurrence_interval_prob}
\title{recurrence_interval}
\usage{
recurrence_interval_prob(Y, R)
}
\arguments{
\item{Y}{years in record}

\item{R}{magnitude ranking}

\item{T}{recurrence interval (note: not an independent input, but a dependent input)}

\item{Prob}{the probability (%) in any given year that a flood of that magnitude will be equaled or exceeded}
}
\value{
the risk in in any given year that a flood of that magnitude will be equaled or exceeded
}
\description{
This function finds the probability in any given year that a flood of that magnitude will be equaled or exceeded
}
\author{
Claudia Flores
}
